pyproject.toml
[project]
name = "job-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fake-useragent>=2.2.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest-playwright>=0.7.2",
]
